+++ CHANGELOG.md
@@ -34,6 +34,11 @@ Breaking changes should be prefixed by `[**BREAKING**]` (without the quotes), to
 
 - Dropped support for end of life Ruby on Rails 7.1. It _may_ still work, we don't test it anymore
 
+### Fixed
+
+- Tasks:
+  - Replace factory names in `federails:copy_factories`
+
 ## [0.7.0] 2025-05-10
 
 ### Added